Remove FutureWarning: Setting an item of incompatible dtype is deprecated (#374)

+++ ctgan/data_transformer.py
@@ -189,7 +189,8 @@ class DataTransformer(object):
 
     def _inverse_transform_continuous(self, column_transform_info, column_data, sigmas, st):
         gm = column_transform_info.transform
-        data = pd.DataFrame(column_data[:, :2], columns=list(gm.get_output_sdtypes()))
+        data = pd.DataFrame(
+            column_data[:, :2], columns=list(gm.get_output_sdtypes())).astype(float)
         data[data.columns[1]] = np.argmax(column_data[:, 1:], axis=1)
         if sigmas is not None:
             selected_normalized_value = np.random.normal(data.iloc[:, 0], sigmas[st])
